backend/builditon/Backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
pdfplumber
sarvamai
google-generativeai
supabase